refactored m3 drawables
